int -> float

+++ jetstream/core/metrics/prometheus.py
@@ -25,7 +25,7 @@ class JetstreamMetricsCollector:
   _id: str = os.getenv("HOSTNAME", shortuuid.uuid())
   server_start_time: int
 
-  def __new__(cls, server_start_time: int):
+  def __new__(cls, server_start_time: float):
     if not hasattr(cls, "instance"):
       cls.instance = super(JetstreamMetricsCollector, cls).__new__(cls)
     cls.server_start_time = server_start_time
